03-Curso_2/05-restapi/restapi03/requirements.txt
aws-cdk-lib==2.144.0
constructs>=10.0.0,<11.0.0
boto3>=1.26